requirements.txt
easyocr
mss
pillow
numpy
tkinter-tooltip
